+++ src/i18n_check/test_frontends/all_checks_fail/test_file.ts
@@ -2,6 +2,7 @@
 //   - "i18n._global.hello_global"
 //   - "i18n._global.hello_global_repeat_value"
 //   - "i18n._global.repeat_key"
+//   - "i18n.test_file.form_button_aria_label"
 //   - i18n.test_file.hello_test_file  // no quotes or backticks
 //   - "i18n.test_file.incorrectly-formatted-key"
 //   - "i18n.test_file.nested_example"