+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 
 ### ✨ Features
 
+- A new aria label check has been added to make sure that i18n keys that end in `_aria_label` do not end in punctuation ([#48](https://github.com/activist-org/i18n-check/issues/48)).
 - The user can now ignore certain regex patterns in their i18n keys within the invalid keys check ([#52](https://github.com/activist-org/i18n-check/issues/52), [#58](https://github.com/activist-org/i18n-check/issues/58)).
 - In the invalid keys check the `--fix` option now uses just the global directories and files to skip to assure that all instances of keys are replaced even if their locations should not be factored into the names of the key ([#59](https://github.com/activist-org/i18n-check/issues/59)).